requirements.txt
asgiref==3.7.2
Django==5.0
django-modeltranslation==0.18.11
django-widget-tweaks==1.5.0
Pillow==10.1.0
sqlparse==0.4.4
typing_extensions==4.9.0
